requirements.txt
fastapi>=0.110.0,<0.113.0
uvicorn[standard]>=0.30.0,<0.31.0
jinja2>=3.1.0,<4.0.0
weasyprint>=61.0,<63.0
pydantic>=2.6.0,<3.0.0
pytest>=7.4.0,<8.0.0
python-dotenv>=1.0.1,<2.0.0
